Refactor: RP-161 컨트롤러를 인바운드 어댑터로 이관
