+++ src/main/java/com/cherrypick/backend/domain/deal/adapter/in/web/RecommenderController.java
@@ -1,6 +1,5 @@
-package com.cherrypick.backend.domain.deal.controller;
+package com.cherrypick.backend.domain.deal.adapter.in.web;
 
-import com.cherrypick.backend.domain.deal.dto.request.DealRequestDTOs;
 import com.cherrypick.backend.domain.deal.service.RecommenderService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
